fix: return clear error message for unique constraints (#138)
* migrate: ignore unique constraints

* chore: return error for unique constraints

Return an error when trying to create a unique constraint, instead of
silently ignoring them. This makes the use of unique constraints the
same both when creating and altering a table. It also makes it
possible to return a hint as to what the user should do (use a unique
index instead).

This commit also moves the test code from the integration_test.go
file to the migrator_emulator_test.go, as those tests are much
faster to execute, and also do not require access to GitHub secrets.
This means that the tests will be executed for each pull request,
including pull requests coming from forks.

* fix: column introspection should return false for unique column

* chore: clarify comment

---------

Co-authored-by: Knut Olav Løite <koloite@gmail.com>