perf: use last_statement for auto-commit
Statements that are executed without a default transaction use auto-commit with the
underlying database/sql driver. Using this allows the driver to set the last_statement=true
option for the statement, which again allows Spanner to optimize the execution of the
statement, as it knows that it is the last statement in the read/write transaction.